Team Projects of Studypod
All are in console windows projects.

First is DLS calculator that is base on dls method of circket. It can calculate reduced overs and winning team .

Second is casino game that is a casino game of random number matching lucky drop wheel.

Login page .

Restaurant Project is new idea that base the user input of amount and we show a menu which things they can buy and they can order it as well. And also, if amount is grater then bill amount then they can order more this, the program automatically show this menu.